Docker y mariadb funcionando

+++ ApiProducto/src/main/kotlin/com/example/apiproducto/repositories/ServiceRepository.kt
@@ -3,4 +3,4 @@ package com.example.apiproducto.repositories
 import com.example.apiproducto.models.Service
 import org.springframework.data.repository.kotlin.CoroutineCrudRepository
 
-interface ServiceRepository : CoroutineCrudRepository<Service, Long>
+interface ServiceRepository : CoroutineCrudRepository<Service, Int>
